requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
Django==5.0.6
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
